remove compiled files.  add failing test for refactor.
Adds domain function to roll the dice.
Adds unimplemented function to calculate the output message for a roll.
