--- H5_Skunk/bin/.gitignore
@@ -1,19 +1,0 @@
-/Dice.class
-/Die.class
-/DieTest.class
-/Player.class
-/SkunkApp.class
-/SkunkDomain.class
-/SkunkUI.class
-/UI.class
-/MockUI.class
-/MockUI_Yes.class
-/MockUI_No.class
-/TestSkunkDomain.class
-/TestMockUI.class
-/TestSkunkDomain2.class
-/TestSkunkDomain3.class
-/TestSkunkDomain4.class
-/TestSkunkDomain5.class
-/TestSkunkDomain6.class
-/TestSkunkDomain7.class
